pyproject.toml
[project]
name = "hashcalc"
description = "Default template for PDM package"
authors = [
    {name = "Marie Roald", email = "marie.roald@nb.no"},
]
dependencies = ["fastapi[standard]>=0.116.1", "uvicorn>=0.35.0", "python-multipart>=0.0.20", "numpy>=2.3.2", "pillow>=11.3.0"]
requires-python = ">=3.13"
readme = "README.md"
license = {text = "GPL-3.0"}
dynamic = ["version"]


[project.optional-dependencies]
pdqhash = [
    "pdqhash>=0.2.8",
]
[build-system]
requires = ["maturin>=1.7,<2.0"]
build-backend = "maturin"

[tool.maturin]
features = ["pyo3/extension-module"]
python-source = "python"
module-name = "hashcalc._native"

[tool.ruff.lint]
select = ["I", "TCH"]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "ruff>=0.12.9",
]


[tool.bumpversion]
current_version = "0.0.1"

parse = "(?P<major>\\d+)\\.(?P<minor>\\d+)\\.(?P<patch>\\d+)"
serialize = ["{major}.{minor}.{patch}"]

search = "\nversion = \"{current_version}\""
replace = "\nversion = \"{new_version}\""

commit = true
message = "Bump version: {current_version} → {new_version}"

tag = true
tag_name = "v{new_version}"
tag_message = "Bump version: {current_version} → {new_version}"

[[tool.bumpversion.files]]
filename = "Cargo.toml"
